Drop DVF center X and Y camera features
These features allow one to ask the device to automatically define a ROI
which is centered in the X or Y axis. This is not needed here because a
machinery for defining the center is already done in higher levels and
we cannot rely on the fact that the beam will always hit the sensor
center. In fact, these PVs are not mapped to DVF methods. Remove them
from the properly list entirely.

This removes the need to translate their references to the new
GC_Center{X,Y} names.