Disregard ffmpeg plugin state during DVF reset
This plugins provides a compressed data stream to be accessed over HTTP.
Standard IOC deploy no longer includes it for beamlines. Since we don't
actually need it, but rather ensure it is always disabled, do not
attempt to connect to its PVs.

It is assumed that accelerator deploy does not enable by default nor
often use it.