Drop usage of DVF camera PixelSize PV
This PV is a specific feature of the camera that reports the current
bits depth given the PixelFormat. For Mono8, it will be Bpp8, while
Mono10 will be Bpp16, and so on. Since this is currently not used, and
it is redundant with the standard ADDriver DataType, drop its usage
entirely.