orbintlk.FIX: fix timing redundancy check
The previous version of the code considered the value of the
RxLockedLtc PV as a whole, without checking the specific bit for that
redundancy device. This used to work when we had only one device per
Fout enabled, because in case of failures, RxLockedLtc would assume the
value 0. For more than one device (AFC Timing or redundancy EVR) in the
same Fout, the checks would verify that RxLockedLtc was not 0, but this
will not happen unless all devices in Fout lose the lock.