orbintlk.FIX: fix bug in redundancy checks
The previous version was considering a redundancy EVR device would not
be shared between AFCTIs: since the intlkr_fouttable dictionary was a
1-to-1 mapping, it would only check the last pair registered as
redundancy, since any previous value would be overwritten. This was
fixed by changing the intlkr_fouttable dictionary into a 1-to-many
mapping and changing the checks to consider combined status of AFCTIs
when a EVR is down.