macreport.ENH: do not ignore wrong shift annotations in the beggining of an analized period
There is a feature that was introduced to ignore wrong shift annotations within 1min (for example, to not indicate user shift problems if the MachShift mode PV changes in value 1min later than scheduled). This commit changes this feature to not ignore wrong shift annotations in the beggining of an analized period. This was causing a wrong indication of transition from "correct shift" to "wrong shift" when the analized period started with a wrong shift and, as a consequence, causing a wrong beam dump annotation.